Refactor Teams structure, update styling, and remove deprecated data
