+++ components/Alumni/Alumni.module.scss
@@ -1,4 +1,4 @@
-@import '../../variables';
+@import '../variables';
 
 .filter {
   display: flex;
