+++ components/ContributionCard/ContributionCard.module.scss
@@ -6,9 +6,17 @@
   min-height: 22rem;
   display: flex;
   flex-direction: column;
-  background-color: #1a1a1a;
+  background-color: rgba(26, 26, 26, 0.6);
+  backdrop-filter: blur(5px);
   border-radius: 10px;
   overflow: hidden;
+  transition: all 0.3s ease;
+
+  &:hover {
+    background-color: rgba(26, 26, 26, 0.8);
+    transform: translateY(-5px);
+    cursor: pointer;
+  }
 
   .top {
     height: 6rem;
@@ -73,6 +81,16 @@
         background-color: #0d0d0d;
         color: #4caf50;
         border: 1px solid #4caf50;
+
+        &.lfx {
+          color: #ff4d4d;
+          border-color: #ff4d4d;
+        }
+
+        &.ospp {
+          color: #ffeb3b;
+          border-color: #ffeb3b;
+        }
       }
     }
