--- components/Teams/Current/Current.module.scss
@@ -1,16 +1,0 @@
-@import '../../variables';
-
-.cardGrp {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
-  gap: 2rem;
-  padding: 2rem;
-
-  @media (max-width: 850px) {
-    grid-template-columns: 1fr 1fr;
-  }
-
-  @media (max-width: 650px) {
-    grid-template-columns: 1fr;
-  }
-}
