--- components/Teams/TeamSelection/TeamsIndex.module.scss
@@ -1,65 +1,0 @@
-.teamSelect {
-  width: 100%;
-  height: 100vh;
-  // overflow: auto;
-
-  background-image: linear-gradient(rgba(0, 0, 0, 0.6), rgba(0, 0, 0, 0.6)),
-    url('/static/images/hero/team-selection-hero.jpg');
-  background-size: cover;
-  background-repeat: no-repeat;
-  background-position: center;
-
-  h1 {
-    font-size: 4rem;
-    font-family: 'Oswald', sans-serif;
-    padding-top: 1.5em;
-    text-align: center;
-  }
-
-  .teamSelList {
-    width: 40%;
-    margin: auto;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: stretch;
-    padding: 3em;
-
-    .link {
-      text-align: center;
-      font-size: 2rem;
-      border: 2px solid white;
-      border-radius: 2px;
-      padding: 0.5em 1em;
-      margin: 1em;
-      &:hover {
-        background-color: rgba(255, 255, 255, 0.164);
-      }
-    }
-    @media (max-width: 1250px) {
-      width: 50%;
-    }
-    @media (max-width: 1000px) {
-      width: 60%;
-    }
-    @media (max-width: 850px) {
-      width: 70%;
-    }
-    @media (max-width: 750px) {
-      width: 70%;
-      .link {
-        font-size: 1.5rem;
-      }
-    }
-    @media (max-width: 575px) {
-      width: 80%;
-      padding: 0.5em;
-    }
-    @media (max-width: 500px) {
-      width: 90%;
-    }
-    @media (max-width: 450px) {
-      width: 95%;
-    }
-  }
-}
