--- data/faculty.js
@@ -1,25 +1,0 @@
-const FacultyInchargeList = [
-  {
-    name: 'Dr. A. S. Rao',
-    imgName: 'Dr. A. S. Rao.jpg',
-    sub: <>Faculty Coordinator</>,
-    linkedInLink: 'https://www.linkedin.com/in/dr-a-s-rao-a82aa620/',
-    githubLink: '',
-  },
-  {
-    name: 'Dr. V. B. Suryawanshi',
-    imgName: 'Dr. V. B. Suryawanshi.jpg',
-    sub: <>Vice-Chairman</>,
-    linkedInLink: 'https://www.linkedin.com/in/dr-vinod-suryawanshi-b02a8515/',
-    githubLink: '',
-  },
-  {
-    name: 'Dr. S. A. Mastud',
-    imgName: 'Dr. S. A. Mastud.jpg',
-    sub: <>Chairman</>,
-    linkedInLink: 'https://www.linkedin.com/in/sachin-mastud-670602206/',
-    githubLink: '',
-  },
-];
-
-export default FacultyInchargeList;
