--- data/flagshipprojects.js
@@ -1,22 +1,0 @@
-const FlagshipProjectsData = [
-  {
-    imgName: 'Avitra.jpg',
-    name: 'AVITRA',
-    sub: "AVITRA is a 5-DOF Serial Manipulator retro-fitted on a holonomic drive platform capable of mapping unknown environments, autonomously navigating known environments and manipulating small objects of various shapes. It's driven by an Intel NUC, which gives the robot the ability to process the information around it with the help of industry-standard software like ROS due to which development and customization to fit various applications becomes more feasible.",
-    githubLink: 'https://github.com/SRA-AVITRA',
-  },
-  {
-    imgName: 'Venom.jpg',
-    name: 'VENOM (Quadruped)',
-    sub: 'Venom is an unmanned four legged robot equipped with RGB-Depth Camera and an ARM Processor for onboard computation. Most of the robots available in the industry make use of wheels for navigation. Whereas a legged vehicle like Venom, will give multiple-terrain mobility, superior to existing wheeled and tracked vehicles. It can also negotiate terrain with minimum of human guidance and intervention.',
-    githubLink: 'https://github.com/chinmaynehate/Venom',
-  },
-  {
-    imgName: 'drone.jpg',
-    name: 'PUSHPAC (Drone)',
-    sub: 'PUSHPAC ( Polymorphic & Unmanned Surveillance Hybrid Platform with Autonomous Capabilities ) is a hybrid unmanned vehicle developed for functioning both in air and underwater. PUSHPAC is capable of extracting data regarding the bed profile of a water body up to a depth of 10 metres, and also performing surveillance and localisation using its vision capabilities, underwater as well as in flight.',
-    githubLink: 'https://github.com/apoorva-21/PUSHPAC',
-  },
-];
-
-export default FlagshipProjectsData;
