+++ data/opensource.js
@@ -13,15 +13,6 @@ const OpenSourceList = [
         githubLink: 'https://github.com/ShaunakKDatar',
       },
       {
-        contributor: 'Prithvi Tambewagh',
-        repoName: 'RTEMS',
-        repoIcon: 'https://avatars.githubusercontent.com/u/156917709?v=4',
-        description: 'Provide SPARC greth network drivers for LWIP',
-        competition: 'GSoC',
-        prLink: 'https://github.com/rkt-1597/GSoC-2025-Final-Report',
-        githubLink: 'https://github.com/rkt-1597',
-      },
-      {
         contributor: 'Suraj Sonawane',
         repoName: 'The Linux Foundation',
         repoIcon: 'https://avatars.githubusercontent.com/u/129578177?v=4',
@@ -83,6 +74,15 @@ const OpenSourceList = [
         competition: 'OSPP',
         prLink:'https://github.com/embox/embox/pull/3755',
         githubLink: 'https://github.com/Shankari02',
+      },
+      {
+        contributor: 'Prithvi Tambewagh',
+        repoName: 'RTEMS',
+        repoIcon: 'https://avatars.githubusercontent.com/u/156917709?v=4',
+        description: 'Provide SPARC greth network drivers for LWIP',
+        competition: 'GSoC',
+        prLink: 'https://github.com/rkt-1597/GSoC-2025-Final-Report',
+        githubLink: 'https://github.com/rkt-1597',
       },
     ],
   },
