--- data/senate.js
@@ -1,230 +1,0 @@
-const SenateList = [
-  {
-    name: 'Atharva Kashalkar',
-    imgName: 'Team/2026/atharva_kashalkar.jpg',
-    sub: (
-      <>
-        General Secretary <br />
-        B.Tech Electronics ' 26
-      </>
-    ),
-    linkedInLink:
-      'https://www.linkedin.com/in/atharva-kashalkar-19598a297?utm_source=share&utm_campaign=share_via&utm_content=profile&utm_medium=android_app',
-    githubLink: 'https://github.com/RapidRoger18',
-  },
-  {
-    name: 'Shankari Anandakrishnan',
-    imgName: 'Team/2026/Shankari.jpg',
-    sub: (
-      <>
-        Jt. General Secretary <br />
-        B.Tech Electronics and Telecommunication ' 26
-      </>
-    ),
-    linkedInLink: 'https://www.linkedin.com/in/shankari-anand/',
-    githubLink: 'https://github.com/Shankari02',
-  },
-  {
-    name: 'Atharva Atre',
-    imgName: 'Team/2026/Atharva Atre.jpg',
-    sub: (
-      <>
-        Jt. General Secretary <br />
-        B.Tech Electronics ' 26
-      </>
-    ),
-    linkedInLink: 'https://www.linkedin.com/in/atharva-atre-817694252/',
-    githubLink: 'https://github.com/AtharvaAtre',
-  },
-  {
-    name: 'Atharv Shailesh Patil',
-    imgName: 'Team/2026/atharva_patil.webp',
-    sub: (
-      <>
-        Electronics head <br />
-        B.Tech Electronics ' 26
-      </>
-    ),
-    linkedInLink: 'https://www.linkedin.com/in/atharv-patil-a43264304/',
-    githubLink: 'https://github.com/Atharv1035',
-  },
-  {
-    name: 'Sarvesh Patki',
-    imgName: 'Team/2026/_storage_emulated_0_DCIM_Camera_Sarvesh_Patki.jpg',
-    sub: (
-      <>
-        Mechanical head <br />
-        B.Tech Mechanical ' 26
-      </>
-    ),
-    linkedInLink: 'https://www.linkedin.com/in/sarvesh-patki-873b10282',
-    githubLink: 'https://github.com/skybirdsgp',
-  },
-  {
-    name: 'Saish Karole',
-    imgName: 'Team/2026/Saish_Karole.jpg',
-    sub: (
-      <>
-        Sponsorship head <br />
-        B.Tech Electronics ' 26
-      </>
-    ),
-    linkedInLink: 'https://www.linkedin.com/in/saish-karole',
-    githubLink: 'https://github.com/nachtspyder04',
-  },
-  {
-    name: 'Aditya Vivekanand',
-    imgName: 'Team/2026/Aditya_Vivekanand.jpeg',
-    sub: (
-      <>
-        Marketing head <br />
-        B.Tech Electronics and Telecommunication ' 26
-      </>
-    ),
-    linkedInLink:
-      'https://www.linkedin.com/in/aditya-vivekanand-929693221?utm_source=share&utm_campaign=share_via&utm_content=profile&utm_medium=ios_app',
-    githubLink: 'https://github.com/adityavivekanand',
-  },
-  {
-    name: 'Anish Mahadevan',
-    imgName: 'Team/2026/Anish_Mahadevan.jpg',
-    sub: (
-      <>
-        Marketing head <br />
-        B.Tech Electronics ' 26
-      </>
-    ),
-    linkedInLink: 'https://www.linkedin.com/in/anish-mahadevan/',
-    githubLink: 'https://github.com/Faulty404',
-  },
-  {
-    name: 'Atharv Dubey',
-    imgName: 'Team/2026/Atharv_Dubey.jpeg',
-    sub: (
-      <>
-        TY Core
-        <br />
-        B.Tech Electronics and Telecommunication ' 26
-      </>
-    ),
-    linkedInLink: 'https://www.linkedin.com/in/atharv-dubey-10103b24a/',
-    githubLink: 'https://github.com/DoobyDoPap/',
-  },
-  {
-    name: 'Abhinav Ananthu',
-    imgName: 'Team/2026/Abhinav_ananthu.jpg',
-    sub: (
-      <>
-        TY Core
-        <br />
-        B.Tech Electronics and Telecommunication ' 26
-      </>
-    ),
-    linkedInLink: 'https://www.linkedin.com/in/abhinav-ananthu-005a23263/',
-    githubLink: 'https://github.com/herculoxz',
-  },
-  {
-    name: 'Aryan Nanda',
-    imgName: 'Team/2026/Aryan Nanda.jpg',
-    sub: (
-      <>
-        TY Core
-        <br />
-        B.Tech Computer Science ' 26
-      </>
-    ),
-    linkedInLink: 'https://www.linkedin.com/in/aryan-nanda-25b32224b/',
-    githubLink: 'https://github.com/AryanNanda17',
-  },
-  {
-    name: 'Purva Sunil Yeshi',
-    imgName: 'Team/2026/Purva_Yeshi.jpg',
-    sub: (
-      <>
-        Core member
-        <br />
-        B.Tech Electronics and Telecommunication ' 26
-      </>
-    ),
-    linkedInLink: 'https://www.linkedin.com/in/purva-yeshi-6491021a1',
-    githubLink: 'https://github.com/purviyeshi',
-  },
-  {
-    name: 'Suraj Sonawane',
-    imgName: 'Team/2026/Suraj Sonawane.jpg',
-    sub: (
-      <>
-        Core member
-        <br />
-        B.Tech Electronics and Telecommunication ' 26
-      </>
-    ),
-    linkedInLink: 'https://www.linkedin.com/in/suraj-sonawane-aba34b231',
-    githubLink: 'https://github.com/SurajSonawane2415',
-  },
-  {
-    name: 'Mahesh Tupe ',
-    imgName: 'Team/2025/Mahesh_Tupe_25.jpg',
-    sub: <>B.Tech Textile ' 25</>,
-    linkedInLink: 'https://www.linkedin.com/in/mahesh-tupe-879393243/',
-    githubLink: 'https://github.com/Asc91',
-  },
-  {
-    name: 'Zain Siddavatam',
-    imgName: 'Team/2025/Zain_Siddavatam_25.jpeg',
-    sub: <>B.Tech Electrical ' 25</>,
-    linkedInLink: 'https://www.linkedin.com/in/zain-siddavatam/',
-    githubLink: 'https://github.com/SuperChamp234',
-  },
-  {
-    name: 'Alqama Shaikh',
-    imgName: 'Team/2025/Alqama_Shaikh_25.jpeg',
-    sub: <>B.Tech Mechanical ' 25</>,
-    linkedInLink: 'https://www.linkedin.com/in/mohd-alqama-shaikh-636587229/',
-    githubLink: 'https://github.com/aPR0T0',
-  },
-  {
-    name: 'Harsh Bhosale',
-    imgName: 'Team/2025/Harsh.jpeg',
-    sub: <>B.Tech Electronics ' 25</>,
-    linkedInLink: 'https://in.linkedin.com/in/hsbh/',
-    githubLink: 'https://github.com/harshbhosale01',
-  },
-  {
-    name: 'Janhavi Deshpande',
-    imgName: 'Team/2025/Janhavi.jpg',
-    sub: <>B.Tech Mechanical ' 25</>,
-    linkedInLink: 'https://in.linkedin.com/in/janhavi-deshpande-a104b8238/',
-    githubLink: 'https://github.com/janhavi1803',
-  },
-  {
-    name: 'Advait Dhamorikar',
-    imgName: 'Team/2025/Advait.jpg',
-    sub: <>B.Tech Electrical ' 25</>,
-    linkedInLink: 'https://in.linkedin.com/in/advait-dhamorikar/',
-    githubLink: 'https://github.com/advait-0',
-  },
-  {
-    name: 'Vaidic Gupta',
-    imgName: 'Team/2025/Vaidic.JPG',
-    sub: <>B.Tech Electrical ' 25</>,
-    linkedInLink: 'https://in.linkedin.com/in/vaidic-gupta-605257230/',
-    githubLink: 'https://github.com/Vaidicgupta',
-  },
-  {
-    name: 'Pratham Deshmukh',
-    imgName: 'Team/2025/Pratham.jpg',
-    sub: <>B.Tech Electrical ' 25</>,
-    linkedInLink: 'https://www.linkedin.com/in/pratham-deshmukh-a027a7227/',
-    githubLink: 'https://github.com/Pratham-Bot',
-  },
-  {
-    name: 'Aryan Karawale',
-    imgName: 'Team/2025/Aryan.jpg',
-    sub: <>B.Tech Electronics & Telecommunication ' 25</>,
-    linkedInLink: 'https://www.linkedin.com/in/aryan-karawale-54b6b7233/',
-    githubLink: 'https://github.com/Aryan-karawale',
-  },
-];
-
-export default SenateList;
