+++ pages/alumni.js
@@ -1,0 +1,15 @@
+import Alumni from '../components/Alumni/Alumni';
+import Head from 'next/head';
+
+const AlumniPage = () => {
+    return (
+        <>
+            <Head>
+                <title>Alumni | SRA VJTI</title>
+            </Head>
+            <Alumni />
+        </>
+    );
+};
+
+export default AlumniPage;
