+++ pages/opensource.js
@@ -1,0 +1,15 @@
+import OpenSource from '../components/Opensource/Opensource';
+import Head from 'next/head';
+
+const OpenSourcePage = () => {
+    return (
+        <>
+            <Head>
+                <title>Open Source | SRA VJTI</title>
+            </Head>
+            <OpenSource />
+        </>
+    );
+};
+
+export default OpenSourcePage;
