--- pages/teams/alumni.js
@@ -1,15 +1,0 @@
-import Alumni from '../../components/Teams/Alumni/Alumni';
-import Head from 'next/head';
-
-const AlumniPage = () => {
-  return (
-    <>
-      <Head>
-        <title>Alumni | SRA VJTI</title>
-      </Head>
-      <Alumni />
-    </>
-  );
-};
-
-export default AlumniPage;
