--- pages/teams/current.js
@@ -1,15 +1,0 @@
-import Current from '../../components/Teams/Current/Current';
-import Head from 'next/head';
-
-const CurrentPage = () => {
-  return (
-    <>
-      <Head>
-        <title>Current Team | SRA VJTI</title>
-      </Head>
-      <Current />
-    </>
-  );
-};
-
-export default CurrentPage;
