--- pages/teams/index.js
@@ -1,15 +1,0 @@
-import TeamsIndex from '../../components/Teams/TeamSelection/TeamsIndex';
-import Head from 'next/head';
-
-const TeamsPage = () => {
-  return (
-    <>
-      <Head>
-        <title>Team | SRA VJTI</title>
-      </Head>
-      <TeamsIndex />
-    </>
-  );
-};
-
-export default TeamsPage;
