--- pages/teams/opensource.js
@@ -1,15 +1,0 @@
-import OpenSource from '../../components/Teams/Opensource/Opensource';
-import Head from 'next/head';
-
-const OpenSourcePage = () => {
-  return (
-    <>
-      <Head>
-        <title>Open Source | SRA VJTI</title>
-      </Head>
-      <OpenSource />
-    </>
-  );
-};
-
-export default OpenSourcePage;
