--- pages/teams/teacherincharge.js
@@ -1,15 +1,0 @@
-import TeacherIncharge from '../../components/Teams/TeacherIncharge/TeacherIncharge';
-import Head from 'next/head';
-
-const TeacherInchargePage = () => {
-  return (
-    <>
-      <Head>
-        <title>Teachers Incharge | SRA VJTI</title>
-      </Head>
-      <TeacherIncharge />
-    </>
-  );
-};
-
-export default TeacherInchargePage;
